+++ Libplanet/Blockchain/Policies/BlockPolicy.cs
@@ -18,6 +18,7 @@ namespace Libplanet.Blockchain.Policies
         private readonly int _maxBlockBytes;
         private readonly int _maxGenesisBytes;
         private readonly Func<Transaction<T>, BlockChain<T>, bool> _doesTransactionFollowPolicy;
+        private readonly HashAlgorithmGetter _hashAlgorithmGetter;
 
         /// <summary>
         /// Creates a <see cref="BlockPolicy{T}"/> with configuring
@@ -46,6 +47,8 @@ namespace Libplanet.Blockchain.Policies
         /// </param>
         /// <param name="canonicalChainComparer">The custom rule to determine which is the canonical
         /// chain.  If omitted, <see cref="TotalDifficultyComparer"/> is used by default.</param>
+        /// <param name="hashAlgorithmGetter">Configures <see cref="GetHashAlgorithm(long)"/>.
+        /// If omitted, SHA-256 is used for every block.</param>
         public BlockPolicy(
             IAction blockAction = null,
             int blockIntervalMilliseconds = 5000,
@@ -55,7 +58,8 @@ namespace Libplanet.Blockchain.Policies
             int maxBlockBytes = 100 * 1024,
             int maxGenesisBytes = 1024 * 1024,
             Func<Transaction<T>, BlockChain<T>, bool> doesTransactionFollowPolicy = null,
-            IComparer<BlockPerception> canonicalChainComparer = null
+            IComparer<BlockPerception> canonicalChainComparer = null,
+            HashAlgorithmGetter hashAlgorithmGetter = null
         )
             : this(
                 blockAction,
@@ -66,7 +70,8 @@ namespace Libplanet.Blockchain.Policies
                 maxBlockBytes,
                 maxGenesisBytes,
                 doesTransactionFollowPolicy,
-                canonicalChainComparer)
+                canonicalChainComparer,
+                hashAlgorithmGetter)
         {
         }
 
@@ -96,6 +101,8 @@ namespace Libplanet.Blockchain.Policies
         /// chain.  If omitted, <see cref="TotalDifficultyComparer"/> (having
         /// <see cref="TotalDifficultyComparer.OutdateAfter"/> configured to triple of
         /// <paramref name="blockInterval"/>) is used by default.</param>
+        /// <param name="hashAlgorithmGetter">Configures <see cref="GetHashAlgorithm(long)"/>.
+        /// If omitted, SHA-256 is used for every block.</param>
         public BlockPolicy(
             IAction blockAction,
             TimeSpan blockInterval,
@@ -105,7 +112,8 @@ namespace Libplanet.Blockchain.Policies
             int maxBlockBytes,
             int maxGenesisBytes,
             Func<Transaction<T>, BlockChain<T>, bool> doesTransactionFollowPolicy = null,
-            IComparer<BlockPerception> canonicalChainComparer = null
+            IComparer<BlockPerception> canonicalChainComparer = null,
+            HashAlgorithmGetter hashAlgorithmGetter = null
         )
         {
             if (blockInterval < TimeSpan.Zero)
@@ -143,6 +151,7 @@ namespace Libplanet.Blockchain.Policies
             _doesTransactionFollowPolicy = doesTransactionFollowPolicy ?? ((_, __) => true);
             CanonicalChainComparer = canonicalChainComparer
                 ?? new TotalDifficultyComparer(blockInterval + blockInterval + blockInterval);
+            _hashAlgorithmGetter = hashAlgorithmGetter ?? (_ => HashAlgorithmType.Of<SHA256>());
         }
 
         /// <inheritdoc/>
@@ -224,6 +233,6 @@ namespace Libplanet.Blockchain.Policies
 
         /// <inheritdoc cref="IBlockPolicy{T}.GetHashAlgorithm(long)"/>
         public HashAlgorithmType GetHashAlgorithm(long index) =>
-            HashAlgorithmType.Of<SHA256>();
+            _hashAlgorithmGetter(index);
     }
 }